api/requirements.txt
accelerate==0.25.0
APScheduler==3.10.4
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
diffusers==0.24.0
Django==5.0
django-apscheduler==0.6.2
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.9.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
zipp==3.17.0
